Rename `hooks` dir to `wpHooks`
